Time: 312 ms (10.79%), Space: 49.2 MB (16.70%) - LeetHub

+++ 322-coin-change/322-coin-change.cpp
@@ -6,7 +6,7 @@ class Solution {
 public:
     int coinChange(vector<int>& coins, int amount) {
         int n=coins.size();
-        vector<vector<long long>>dp(n+1,vector<long long>(amount+1));
+        vector<vector<int>>dp(n+1,vector<int>(amount+1));
         for(int i=0;i<=n;i++) {
             for(int j=0;j<=amount;j++) {
                 if(j==0) {
